requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
click==8.0.3
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.0.1
